refactor: reset password request
